+++ packages/fields/src/fields/page-title/view.tsx
@@ -1,7 +1,6 @@
 import { __ } from '@wordpress/i18n';
 import { useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
-import type { Settings } from '@wordpress/core-data';
 import { privateApis as componentsPrivateApis } from '@wordpress/components';
 import type { CommonPost } from '../../types';
 import { BaseTitleView } from '../title/view';
@@ -11,10 +10,7 @@ const { Badge: WCBadge } = unlock( componentsPrivateApis );
 export default function PageTitleView( { item }: { item: CommonPost } ) {
 	const { frontPageId, postsPageId } = useSelect( ( select ) => {
 		const { getEntityRecord } = select( coreStore );
-		const siteSettings = getEntityRecord(
-			'root',
-			'site'
-		) as Partial< Settings >;
+		const siteSettings = getEntityRecord( 'root', 'site' );
 		return {
 			frontPageId: siteSettings?.page_on_front,
 			postsPageId: siteSettings?.page_for_posts,